requirements.txt
python-dotenv~=0.14.0
pandas~=1.1.3
praw~=7.1.0
requests~=2.24.0
geopy~=2.0.0
discord~=1.0.1